Explicitly stated destination directory, readlink doesn't show one.
This is meant to copy the emacs-${VERSION} to
emacs-${VERSION}-${XTOOLKIT} to support multiple toolkit targets.

We use readlink to find out what version of emacs got stored at the
destination, and pick up the version for free from the created
symlink. The X toolkit was selected during part of the configure phase
from one of gtk2/gtk3/lucid/motif.

readlink (at least under Bash on Linux) merely spits out a filename
and doesn't provide a pathname along with that, so we specify a
destination path to send the file to.

You do still have to make sure your emacs symlink is pointing to the
version you actually want if you have multiple x-toolkit targets
stored there, as the emacs binary is replaced each time you install.