Lets user explicitly know we're running autogen.sh for them.
The comment isn't entirely correct (running autotools), but someone
with experience will know what happened anyhow. The command that's
actually run is the autogen.sh script at the top of the source tree,
which will execute other programs from the autotools suite.

The reason for this message is to hopefully explain the next line that
may turn up if there are no autotools installed. autogen.sh will fail
to complete in this instance, so it's best to leave the user with no
doubt of what happened.